requirements.txt
SQLAlchemy==1.3.16
tweepy==3.8.0
nltk>=3.6.5
spacy==2.2.4
https://github.com/explosion/spacy-models/releases/download/es_core_news_md-2.2.5/es_core_news_md-2.2.5.tar.gz
python-dotenv==0.13.0
pyinstaller==4.5.1
